refactor: написан класс с обработчиками всех исключений, обновлены отдельно обработчики MethodArgumentNotValidExcption, обработчик остальных Throwable fix: исправлена ошибка с работой методов FilmController
