+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -2,8 +2,6 @@ package ru.yandex.practicum.filmorate.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
@@ -14,7 +12,6 @@ import javax.validation.constraints.Positive;
 import java.util.List;
 
 @Slf4j
-@Validated
 @RestController
 @RequestMapping("films")
 @RequiredArgsConstructor
@@ -55,14 +52,9 @@ public class FilmController {
     }
 
 
+    @Validated
     @GetMapping("/popular")
     private List<Film> getMostPopular(@RequestParam(defaultValue = "10") @Positive int count) {
         return service.getMostPopular(count);
-    }
-
-    @ExceptionHandler
-    public ResponseEntity<String> illegalArgumentExceptionHandler(IllegalArgumentException e){
-        log.warn("Invalid argument {}", e.getMessage());
-        return new ResponseEntity<>("Invalid argument " + e.getMessage(), HttpStatus.BAD_REQUEST);
     }
 }
